Move RepoTrackerBanner into its own folder
